lab1-worker/requirements.txt
pymongo==4.10.1
python-dotenv==1.0.1